# Block & array hybrid storage