… Tests/test_core_resources.py
… Tests/test_image.py
… Tests/test_image_resample.py
… docs/releasenotes/4.3.0.rst