@homm 2017-09-18
Finally all green and almost done