@homm 2017-09-22
I learned a lot about linux memory model this days. For example, there are two mechanisms in Linux for allocation: [mmap](https://en.wikipedia.org/wiki/Mmap) and [sbrk](https://en.wikipedia.org/wiki/Sbrk).

And finally I found something [really interesting](https://github.com/lattera/glibc/blob/master/malloc/malloc.c#L995) in libc sources.


> M_MMAP_THRESHOLD
Using mmap segregates relatively large chunks of memory so that
  they can be individually obtained and released from the host
  system. A request serviced through mmap is never reused by any
  other request (at least not directly; the system may just so
  happen to remap successive requests to the same locations).

So, as discovered above, in some conditions libc doesn't use mmap for blocks up to 32 megabytes and  can reuse them without returning to the system.

Still investigating.