@homm 2017-09-22
Ok, looks like I finally understood everything.

For the [list of libc malloc variables](https://www.gnu.org/software/libc/manual/html_node/Memory-Allocation-Tunables.html) there are only two significant for our case: `mmap_threshold` and `trim_threshold`.

Libc wouldn't return freed memory and would reuse it if two conditions are met:
1) memory was allocated with sbrk, not mmap
2) memory was not released (or trimmed)

`mmap_threshold` is a minimal size of the block which will be allocated using mmap (or, in other words, a maximum size of the block which will be allocated using sbrk)
`trim_threshold` is a minimum size of free memory allocated with sbrk, after which libc will actually release the memory.

By default, both variables are set to 128Kb, but as said in documentation, they are dynamically adjustable. I found only one place in the sources where this happens: [in `__libc_free`](https://github.com/lattera/glibc/blob/master/malloc/malloc.c#L2905).

So, `mmap_threshold` always has a size of the largest freed mmap block. And `trim_threshold` has double size. And this is the most important thing: once an application allocated and freed large enough block of memory (but not larger than `DEFAULT_MMAP_THRESHOLD_MAX`, which is 32Mb on 64-bit systems), it will not release up to two sizes of the large block and will reuse it when possible.

So, let's check this:

A. master, -s2560x1639 (array storage). As you can see, result doesn't match with original comment. This is because there are no scale tests. Indeed, the scale tests bumped `trim_threshold` in 2560x1639→2048x1310 scaling, because 2048x1310 had block storage.
B. master, -s2560x1639 (array storage). Ok, let's try bump  `trim_threshold` with something else, for example 'load' operations.
C. master, -s2560x1639 (array storage). And to be sure, let's try to bump the threshold with simple python code before the tests (in pillow-perf/testsuite/cases/pillow.py):
```python
_tmp = b'0123456789' * int(30*1024*1024*0.1)
_tmp = 0
```
D. master, -s2560x1639 (array storage). How about to set local environment variables manual? `MALLOC_MMAP_THRESHOLD_=33554432 MALLOC_TRIM_THRESHOLD_=67108864`

```
Convert
                      A                B                C                D
  RGB to L        1938.82 Mpx/s    1930.94 Mpx/s    1862.29 Mpx/s    1772.63 Mpx/s
  RGBA to LA      1341.36 Mpx/s    1382.56 Mpx/s    1340.54 Mpx/s    1352.18 Mpx/s
  RGBa to RGBA     749.55 Mpx/s    1360.86 Mpx/s    1364.98 Mpx/s    1359.18 Mpx/s
  RGBA to RGBa    1331.11 Mpx/s    1332.83 Mpx/s    1333.33 Mpx/s    1340.95 Mpx/s

Composition
                      A                B                C                D
  Composition      438.39 Mpx/s     805.80 Mpx/s     804.25 Mpx/s     805.95 Mpx/s

Rotate 
                      A                B                C                D
  Flop             789.88 Mpx/s    1353.53 Mpx/s    1343.92 Mpx/s    1357.92 Mpx/s
  Flip             807.68 Mpx/s    1714.76 Mpx/s    1746.24 Mpx/s    1701.50 Mpx/s
  Rotate 90        391.84 Mpx/s     508.53 Mpx/s     510.31 Mpx/s     512.75 Mpx/s
  Rotate 180       713.68 Mpx/s    1307.09 Mpx/s    1314.51 Mpx/s    1307.48 Mpx/s
  Rotate 270       379.61 Mpx/s     500.03 Mpx/s     495.14 Mpx/s     498.19 Mpx/s
  Transpose        389.07 Mpx/s     520.25 Mpx/s     519.09 Mpx/s     518.12 Mpx/s
  Transverse       397.34 Mpx/s     523.96 Mpx/s     512.26 Mpx/s     526.59 Mpx/s

                      A                B                C                D
real                2.073 s          2.569 s          1.538 s          1.495 s
user                1.516 s          2.524 s          1.508 s          1.468 s
sys                 0.556 s          0.048 s          0.032 s          0.024 s
```

All of three methods work for master and array storage.

to be continued...