@homm 2017-09-22
What about block & array hybrid storage. I choose the block size 1Mb because I think it is large enough to avoid too frequent malloc/free calls and small enough to be allocated even with high memory fragmentation.

C and D cases also bump `trim_threshold` and fix performance for hybrid storage. But no other operations bumps threshold anymore, because largest freed block for now is largest pillow's block and it is 1Mb.

Possible solutions:

1. Leave it as is. Indeed, the problem only related to single memory manger and it's defaults. We are adding something like our own configurable memory manager with more explicit tools to improve performance.
2. Increase default block size to old value: 16Mb. This will bump `trim_threshold` for large images automatically, as before. This will improve compatibility, but user will have to change PILLOW_BLOCK_SIZE to smallest values in the most cases if he decides to manage memory manually.