@wiredfool 2017-09-28
So the invariants here are:

1) Images don't require any more storage then they previously did, except for potentially a handful of bytes due to alignment issues. 
2) Large images are allocating in blocks that are similarly (order of magnitude) sized to the previous chunk size, where pre-patch they would be made of one malloc per line. 
3) Small images will still go in one chunk.
4) The only case where this will use more memory is that it will hold on to a bunch of memory from recently freed images. The high water mark should be the same.