@homm 2017-09-30
> except for potentially a handful of bytes due to alignment issues.

If you mean `arena->alignment`, it is 1 by default and doesn't affect any memory pointers.

> it will hold on to a bunch of memory from recently freed images

I believe for long-living applications where a big enough image (4Mpx) was opened at least once, memory consumption will be the same including held on memory.