@homm 2017-09-30
For some reason one of the docker's builds sometimes fails.