@hugovk 2017-09-30
We're tracking the Docker problem in https://github.com/python-pillow/Pillow/issues/2758.